requirements.txt
aiogram==2.25.1
celery==4.4
Django==3.1.2
django-celery-beat==2.0
django-celery-results==2.0
django-filter==21.1
djangorestframework==3.14.0
lxml==4.9.2
redis==4.5.4
requests==2.28.2
selenium==4.8.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2

